+++ src/Form/LegacyConsumer/templates/base.html.php
@@ -1,9 +1,9 @@
-<?php /** @var Give\Framework\FieldsAPI\Field $field */ ?>
+<?php /** @var Give\Framework\FieldsAPI\Field|\Give\Framework\FieldsAPI\Text $field */ ?>
 <?php /** @var string $typeAttribute */ ?>
 <input
 	type="<?php echo $typeAttribute; ?>"
 	name="<?php echo $field->getName(); ?>"
-	placeholder="<?php echo $field->getLabel(); ?>"
+	placeholder="<?php echo $field->getPlaceholder(); ?>"
 	id="give-<?php echo $field->getName(); ?>"
 	value="<?php echo $field->getDefaultValue(); ?>"
 	<?php echo $field->isRequired() ? 'required' : ''; ?>