+++ src/Form/LegacyConsumer/templates/select.html.php
@@ -1,20 +1,27 @@
 <?php /** @var Give\Framework\FieldsAPI\Select $field */ ?>
 <select
-	name="give_<?php echo $field->getName(); ?>"
+	name="<?php echo $field->getName(); ?><?php echo $field->getAllowMultiple() ? '[]' : ''; ?>"
 	id="give-<?php echo $field->getName(); ?>"
 	<?php echo $field->getAllowMultiple() ? 'multiple' : ''; ?>
 	<?php echo $field->isRequired() ? 'required' : ''; ?>
 	<?php echo $field->isReadOnly() ? 'readonly' : ''; ?>
 >
-	<?php if ( $placeholder = $field->getPlaceholder() ) : ?>
-		<option value=""><?php echo $placeholder; ?></option>
-	<?php endif; ?>
+	<?php
+	if ( $placeholder = $field->getPlaceholder() ) {
+		printf(
+			'<option value="" %2$s %3$s>%1$s</option>',
+			$placeholder,
+			$field->isRequired() ? 'disabled' : '',
+			$field->getDefaultValue() ? '' : 'selected'
+		);
+	}
+	?>
 	<?php foreach ( $field->getOptions() as $option ) : ?>
+		<?php $value = esc_attr( $option->getValue() ); ?>
 		<?php $label = $option->getLabel(); ?>
-		<?php $value = $option->getValue(); ?>
 		<?php $default = $field->getDefaultValue() === $option->getValue(); ?>
 		<option
-			<?php echo $label ? "value={$value}" : ''; ?>
+			<?php echo $label ? "value=\"$value\"" : ''; ?>
 			<?php echo $default ? 'selected' : ''; ?>
 		>
 			<?php echo $label ?: $value; ?>