+++ src/Form/LegacyConsumer/templates/textarea.html.php
@@ -1,6 +1,6 @@
 <?php /** @var Give\Framework\FieldsAPI\Textarea $field */ ?>
 <textarea
-	name="give_<?php echo $field->getName(); ?>"
+	name="<?php echo $field->getName(); ?>"
 	id="give-<?php echo $field->getName(); ?>"
 	<?php echo $field->isRequired() ? 'required' : ''; ?>
 	<?php echo $field->isReadOnly() ? 'readonly' : ''; ?>